% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS-calc-BE.R
\name{calc_be}
\alias{calc_be}
\title{Calculation of BE statistics}
\usage{
calc_be(
  data,
  n_trials = 50,
  subj_min = 5,
  subj_max = 50,
  subj_step = 5,
  seed = -17,
  ci_lcut = 0.8,
  ci_ucut = 1.25
)
}
\arguments{
\item{data}{A dataframe or tibble  with these columns:
subject id, time, concentrations, period, drug.}

\item{n_trials}{Integer coding the number of random trials per sample size.}

\item{subj_min}{Integer defining the minimum number of subjects per trial}

\item{subj_max}{Integer defining the maximum number of subjects per trial}

\item{subj_step}{Integer defining the step size between \code{subj_min} and
\code{subj_max}, so that the program will test sample sizes
$N = \{subj_min, subj_min + subj_step, subj_min + 2*subj_step,...,subj_max\}$.}

\item{seed}{Random number seed.}

\item{ci_lcut}{Lower limit of the confidence interval.}

\item{ci_ucut}{Upper limit of the confidence interval.}
}
\value{
A list of lists: \verb{[[1:n_samp]][[1:n_trials]][["par", "cross", "rep"]][[1,2]]},
where \code{n_samp} is the number of sample sizes tested, e.g.,
(\code{subj_max} - \code{subj-min}) / \code{subj_step}.  The first item in the
design is auc, and the second is cmax.  For example the comparison of auc using
the third trial of the second sample size for the replicated design would be
\code{x[[2]][[3]]$rep[[1]]}.
}
\description{
Generates bioequivalence statistics
}
\details{
Generates BE statistics for parallel, cross-over and full replication designs.
From simulated data with 1000 subjects per drug, the function first calculates
area under the concentration-time curve (AUC) and maximum concentration (Cmax)
for each simulated profiles. Periods are defined as >1 for each replicated
administration (e.g., when intra-occasion varibility is simulated.)
Linear regression (\code{stats::lm()}) is used on \code{log10(auc)} and\code{log10(cmax)}
for parallel and cross-over designs, and mixed effect modeling (\code{lmerTest::lmer()})
is used for the replicated design.
}
\author{
Michael Neely

Abdullah Hamadeh
}
