% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{createVirtualPopulation}
\alias{createVirtualPopulation}
\title{createVirtualPopulation}
\usage{
createVirtualPopulation(
  inferredDistribution,
  proportionOfFemales = 50,
  numberOfVirtualIndividuals = 100,
  clusters,
  demographyRanges
)
}
\arguments{
\item{inferredDistribution}{The inferred distribution.}

\item{proportionOfFemales}{Proportion of females in the virtual population in percent}

\item{numberOfVirtualIndividuals}{The number of virtual individuals to create.}

\item{clusters}{List of clusters.}

\item{demographyRanges}{The demography ranges.}
}
\description{
Create a virtual population.
}
