% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS-extract-BE.R
\name{extract_be}
\alias{extract_be}
\title{Extract  BE statistics}
\usage{
extract_be(lmer_res, ci = 0.9, ci_lcut = 0.8, ci_ucut = 1.25)
}
\arguments{
\item{lmer_res}{The output of (\code{\link{calc_BE}}).}

\item{ci}{Confidence interval for geometric mean ratio, default is 0.9 (90\%)
per FDA guidance.}

\item{ci_lcut}{Lower confidence interval cutoff for geometric mean AUC or Cmax to be
considered BE, default 0.8 per FDA guidance.}

\item{ci_ucut}{Upper confidence interval cutoff for geometric mean AUC or Cmax to be
considered BE, default 1.25 per FDA guidance.}
}
\value{
A list of lists: [[1:n_samp]][[1:\code{n_trials}]][[$par, $cross,
$rep]][[1,2]], where n_samp is the number of sample sizes tested, e.g.,
(\code{subj_max} - \code{subj-min}) / \code{subj_step}.  The first item in the
design is auc, and the second is cmax.  For example the comparison of auc using
the third trial of the second sample size for the replicated design would be
x[[2]][[3]]$rep[[1]].
}
\description{
Extracts BE statistics from results of (\code{\link{calc_BE}}).
}
\seealso{
\code{\link{extractBE}}, \code{\link{plotBE}}
}
\author{
Michael Neely

Abdullah Hamadeh
}
