% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS-extract-BE.R
\name{get_summary}
\alias{get_summary}
\title{get_summary}
\usage{
get_summary(lmer_res, summary_settings, design, index)
}
\arguments{
\item{lmer_res}{The output of (\code{calc_BE()}).}

\item{summary_settings}{A list of settings for the summary function.}

\item{design}{The design of the study, either \code{"parallel"}, \code{"crossover"}, or \code{"replicate"}.}

\item{index}{The index of the design, either 1 or 2.}
}
\value{
A tibble with the summary statistics.
}
\description{
get_summary
}
\author{
Michael Neely

Abdullah Hamadeh
}
