% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{getClusters}
\alias{getClusters}
\title{getClusters}
\usage{
getClusters(inferredDistribution, numberOfClusters)
}
\arguments{
\item{inferredDistribution}{The inferred distribution.}

\item{numberOfClusters}{The number of clusters.}
}
\description{
Get clusters.
}
