% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-main.R
\name{getIndividualCharacteristics}
\alias{getIndividualCharacteristics}
\title{getIndividualCharacteristics}
\usage{
getIndividualCharacteristics(studyPopulationData)
}
\arguments{
\item{studyPopulationData}{A data.frame}
}
\value{
A list of individuals with characteristics
}
\description{
A function to create a list of individuals with \code{ospsuite} characteristics
}
