% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getParameterBounds}
\alias{getParameterBounds}
\title{Retrieve parameter bounds from a list of VBEParameter objects.}
\usage{
getParameterBounds(parameterList)
}
\arguments{
\item{parameterList}{List of \code{VBEParameter} objects.}
}
\value{
A list with two elements: lower bounds and upper bounds.
}
\description{
Retrieve parameter bounds from a list of VBEParameter objects.
}
