% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getParameterPathsInReferenceSimulation}
\alias{getParameterPathsInReferenceSimulation}
\title{Retrieve parameter paths in the reference simulation from a list of VBEParameter objects.}
\usage{
getParameterPathsInReferenceSimulation(parameterList)
}
\arguments{
\item{parameterList}{List of \code{VBEParameter} objects.}
}
\value{
A character vector of parameter paths in the reference simulation.
}
\description{
Retrieve parameter paths in the reference simulation from a list of VBEParameter objects.
}
