% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getParameterPathsInTestSimulation}
\alias{getParameterPathsInTestSimulation}
\title{Retrieve parameter paths in the test simulation from a list of VBEParameter objects.}
\usage{
getParameterPathsInTestSimulation(parameterList)
}
\arguments{
\item{parameterList}{List of \code{VBEParameter} objects.}
}
\value{
A character vector of parameter paths in the test simulation.
}
\description{
Retrieve parameter paths in the test simulation from a list of VBEParameter objects.
}
