% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{getReferenceGoodnessOfFitPlot}
\alias{getReferenceGoodnessOfFitPlot}
\title{getReferenceGoodnessOfFitPlot}
\usage{
getReferenceGoodnessOfFitPlot(
  refAndTestSimulationsInVirtualPopulation,
  pkDataFilePath,
  referenceDrugName = "Reference",
  yDimension = NULL,
  yUnit = NULL,
  timeUnit = ospsuite::ospUnits$Time$min
)
}
\arguments{
\item{refAndTestSimulationsInVirtualPopulation}{A data frame containing the simulated plasma concentrations for the reference and test drugs.}

\item{pkDataFilePath}{The path to the observed plasma concentrations data.}

\item{referenceDrugName}{The name of the reference drug.}

\item{yDimension}{The dimension of the y-axis.}

\item{yUnit}{The unit of the y-axis.}

\item{timeUnit}{The unit of the time axis.}
}
\description{
Generates plots of the goodness of fit for the reference drug.
}
