% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{getReferenceVBEPlot}
\alias{getReferenceVBEPlot}
\title{getReferenceVBEPlot}
\usage{
getReferenceVBEPlot(
  refAndTestSimulationsInVirtualPopulation,
  referenceDrugName = "Reference",
  testDrugName = "Test",
  yDimension = NULL,
  yUnit = NULL,
  timeUnit = ospsuite::ospUnits$Time$min
)
}
\arguments{
\item{refAndTestSimulationsInVirtualPopulation}{A data frame containing the simulated plasma concentrations for the reference and test drugs.}

\item{referenceDrugName}{The name of the reference drug.}

\item{testDrugName}{The name of the test drug.}

\item{yDimension}{The dimension of the y-axis.}

\item{yUnit}{The unit of the y-axis.}

\item{timeUnit}{The unit of the time axis.}
}
\description{
Generates plots of the virtual bioequivalence for the reference drug.
}
