% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{globalVariableNames}
\alias{globalVariableNames}
\title{globalVariableNames}
\format{
An object of class \code{character} of length 14.
}
\usage{
globalVariableNames
}
\description{
Remove check warning for the variables listed below
Variables called withing dplyr and tidyr functions
}
\keyword{internal}
