% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-utilities.R
\name{log10Safe}
\alias{log10Safe}
\title{Safe Base-10 Logarithm Function}
\usage{
log10Safe(x)
}
\arguments{
\item{x}{Numeric vector. The values for which the base-10 logarithm is computed.}
}
\value{
A numeric vector of base-10 logarithm values, where non-positive inputs are replaced with a small positive threshold.
}
\description{
Computes the base-10 logarithm of a numeric input using \code{logSafe()}.
}
\examples{
log10Safe(c(1, 10, 0, -5))

}
