% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-utilities.R
\name{logSafe}
\alias{logSafe}
\title{Safe Logarithm Function}
\usage{
logSafe(x, base = exp(1), min_value = .Machine$double.xmin)
}
\arguments{
\item{x}{Numeric vector. The values for which the logarithm is computed.}

\item{base}{Numeric. The base of the logarithm. Default is the natural logarithm (\code{exp(1)}).}

\item{min_value}{Numeric. A small positive threshold to replace non-positive values. Default is \code{.Machine$double.xmin}.}
}
\value{
A numeric vector of logarithm values, where non-positive inputs are replaced with \code{min_value} before computing the log.
}
\description{
Computes the logarithm of a numeric input, replacing non-positive values with a small positive threshold to avoid errors.
}
\examples{
logSafe(c(1, 10, 0, -5))
logSafe(c(0.1, 0.5, 0), base = 2)

}
