% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-main.R
\name{NPODRunSettings}
\alias{NPODRunSettings}
\title{NPODRunSettings}
\description{
An NPODRunSettings object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{noise}}{A logical to define if noise is present}

\item{\code{model}}{A model object}

\item{\code{ncycles}}{The number of cycles}

\item{\code{MaxIter}}{The maximum number of iterations}

\item{\code{theta_F}}{The theta_F value}

\item{\code{theta_d}}{The theta_d value}

\item{\code{size_theta0}}{The size of theta0}

\item{\code{cache_folder_name}}{The name of the cache folder}

\item{\code{npod_cache}}{The cache object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NPODRunSettings-new}{\code{NPODRunSettings$new()}}
\item \href{#method-NPODRunSettings-clone}{\code{NPODRunSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPODRunSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-NPODRunSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NPODRunSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPODRunSettings$new(
  noise = FALSE,
  model = NULL,
  ncycles = Inf,
  MaxIter = 10,
  theta_F = 0.1,
  theta_d = 0.001,
  size_theta0 = NULL,
  cache_folder_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{noise}}{A logical to define if noise is present}

\item{\code{model}}{A model object}

\item{\code{ncycles}}{The number of cycles}

\item{\code{MaxIter}}{The maximum number of iterations}

\item{\code{theta_F}}{The theta_F value}

\item{\code{theta_d}}{The theta_d value}

\item{\code{size_theta0}}{The size of theta0}

\item{\code{cache_folder_name}}{The name of the cache folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPODRunSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NPODRunSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPODRunSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
