% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotClinicalTrialSimulationResults}
\alias{plotClinicalTrialSimulationResults}
\title{plotClinicalTrialSimulationResults}
\usage{
plotClinicalTrialSimulationResults(ctsResults, addPlotHeaders = TRUE)
}
\arguments{
\item{ctsResults}{The results of the clinical trial simulation.}

\item{addPlotHeaders}{Whether to add plot headers.}
}
\description{
Generates plots of the clinical trial simulation results.
}
