% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotVBEBands}
\alias{plotVBEBands}
\title{plotVBEBands}
\usage{
plotVBEBands(ctsResults)
}
\arguments{
\item{ctsResults}{The results of the clinical trial simulation.}
}
\description{
Generates plots of the virtual bioequivalence bands.
}
