% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotVBEProbability}
\alias{plotVBEProbability}
\title{plotVBEProbability}
\usage{
plotVBEProbability(ctsResults)
}
\arguments{
\item{ctsResults}{The results of the clinical trial simulation.}
}
\description{
Generates plots of the probability of bioequivalence.
}
