% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-utilities.R
\name{posterior}
\alias{posterior}
\title{Compute Posterior Probabilities}
\usage{
posterior(ans)
}
\arguments{
\item{ans}{A list containing \code{PSI}, a matrix of support points (rows: subjects, columns: support points), and \code{w}, a vector of weights.}
}
\value{
A matrix of posterior probabilities for each subject and support point.
}
\description{
Computes posterior probabilities given estimated support points and weights.
}
\examples{
ans <- list(PSI = matrix(runif(9), 3, 3), w = c(0.2, 0.5, 0.3))
posterior(ans)

}
