% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{runClinicalTrialSimulation}
\alias{runClinicalTrialSimulation}
\title{runClinicalTrialSimulation}
\usage{
runClinicalTrialSimulation(
  virtualPopulationSimulationResults,
  n_trials = 50,
  subj_min = 5,
  subj_max = 50,
  subj_step = 5
)
}
\arguments{
\item{virtualPopulationSimulationResults}{The results of the virtual population simulation.}

\item{n_trials}{The number of trials to run.}

\item{subj_min}{The minimum number of subjects to include in the trial.}

\item{subj_max}{The maximum number of subjects to include in the trial.}

\item{subj_step}{The step size between \code{subj_min} and \code{subj_max}.}
}
\description{
Run a clinical trial simulation.
}
