% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{runDistributionInference}
\alias{runDistributionInference}
\title{runDistributionInference}
\usage{
runDistributionInference(
  method = "NPOD",
  referenceSimulationFilePath,
  outputPath,
  pkDataFilePath,
  studyPopulationDataFilePath = NULL,
  cofactorPaths = NULL,
  inferenceParameters = NULL,
  numberOfIterations = 10,
  useLogNormalLikelihood = NULL,
  initialGridSize = NULL,
  saveResultsPath = NULL,
  cacheFolder = NULL
)
}
\arguments{
\item{method}{The method to use for distribution inference. Currently only \code{"NPOD"} is supported.}

\item{referenceSimulationFilePath}{The path to the reference simulation file.}

\item{outputPath}{The path to the output directory.}

\item{pkDataFilePath}{The path to the PK data file.}

\item{studyPopulationDataFilePath}{The path to the study population data file.}

\item{cofactorPaths}{The paths to the cofactor files.}

\item{inferenceParameters}{The parameters to infer.}

\item{numberOfIterations}{The number of iterations to run.}

\item{useLogNormalLikelihood}{Whether to use a log-normal likelihood.}

\item{initialGridSize}{The initial grid size.}

\item{saveResultsPath}{The path to save the results.}

\item{cacheFolder}{The folder to use for caching.}
}
\description{
Run distribution inference
}
