% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-output.R
\name{SimOutput}
\alias{SimOutput}
\title{SimOutput}
\description{
A class to store simulation output information.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the simulation output}

\item{\code{path}}{Path to the simulation output}

\item{\code{dimension}}{Dimension of the simulation output}

\item{\code{displayUnit}}{Display unit of the simulation output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimOutput-new}{\code{SimOutput$new()}}
\item \href{#method-SimOutput-clone}{\code{SimOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimOutput-new}{}}}
\subsection{Method \code{new()}}{
Initialize the simulation output object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOutput$new(name, path, dimension, displayUnit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the simulation output}

\item{\code{path}}{Path to the simulation output}

\item{\code{dimension}}{Dimension of the simulation output}

\item{\code{displayUnit}}{Display unit of the simulation output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
