% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{simulateVirtualPopulation}
\alias{simulateVirtualPopulation}
\title{simulateVirtualPopulation}
\usage{
simulateVirtualPopulation(
  referenceSimulationFilePath,
  testSimulationFilePath,
  referencePopulationDataframe,
  testPopulationDataframe,
  outputPath,
  startTime,
  endTime,
  resolutionPtsMin
)
}
\arguments{
\item{referenceSimulationFilePath}{The path to the reference simulation file.}

\item{testSimulationFilePath}{The path to the test simulation file.}

\item{referencePopulationDataframe}{The reference population dataframe.}

\item{testPopulationDataframe}{The test population dataframe.}

\item{outputPath}{The output path.}

\item{startTime}{The start time of the simulation.}

\item{endTime}{The end time of the simulation.}

\item{resolutionPtsMin}{The resolution of the simulation.}
}
\description{
Simulate a virtual population.
}
