% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS-calc-BE.R
\name{this_samp}
\alias{this_samp}
\title{this_samp}
\usage{
this_samp(trialNumber, pkParameterData, nSubjects)
}
\arguments{
\item{trialNumber}{The number of the current trial.}

\item{pkParameterData}{The pharmacokinetic parameter data.}

\item{nSubjects}{The number of subjects to sample.}
}
\description{
create subset with n random individuals by drug for parallel design
}
\keyword{internal}
