% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{VBEParameter}
\alias{VBEParameter}
\title{VBEParameter Class}
\description{
VBEParameter Class

VBEParameter Class
}
\details{
An R6 class for handling Virtual Bioequivalence (VBE) parameters.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pathInReferenceSimulation}}{Character. Path to the parameter in the reference simulation.}

\item{\code{pathInTestSimulation}}{Character. Path to the parameter in the test simulation.}

\item{\code{displayName}}{Character. Display name of the parameter.}

\item{\code{dimension}}{Character. Dimension of the parameter.}

\item{\code{unit}}{Character. Unit of the parameter.}

\item{\code{lowerBound}}{Numeric. Lower bound for the parameter.}

\item{\code{upperBound}}{Numeric. Upper bound for the parameter.}

\item{\code{logIncrement}}{Logical. Whether the parameter is incremented logarithmically.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VBEParameter-new}{\code{VBEParameter$new()}}
\item \href{#method-VBEParameter-clone}{\code{VBEParameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VBEParameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-VBEParameter-new}{}}}
\subsection{Method \code{new()}}{
Initialize a VBEParameter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBEParameter$new(
  pathInReferenceSimulation,
  pathInTestSimulation,
  displayName = NULL,
  dimension,
  unit,
  lowerBound = 0.1,
  upperBound = 10,
  logIncrement = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathInReferenceSimulation}}{Character. Path to the parameter in the reference simulation.}

\item{\code{pathInTestSimulation}}{Character. Path to the parameter in the test simulation.}

\item{\code{displayName}}{Character. Display name of the parameter (defaults to \code{pathInReferenceSimulation}).}

\item{\code{dimension}}{Character. Dimension of the parameter.}

\item{\code{unit}}{Character. Unit of the parameter.}

\item{\code{lowerBound}}{Numeric. Lower bound for the parameter (default: 1e-1).}

\item{\code{upperBound}}{Numeric. Upper bound for the parameter (default: 1e1).}

\item{\code{logIncrement}}{Logical. Whether the parameter is incremented logarithmically (default: TRUE).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VBEParameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VBEParameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBEParameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
