% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPOD-utilities.R
\name{getWeightedPoints}
\alias{getWeightedPoints}
\title{Compute Weighted Support Points}
\usage{
getWeightedPoints(res, params = c("Age", "Weight", "Height"))
}
\arguments{
\item{res}{A list containing \code{theta} (a matrix of support points), \code{w} (weights), and \code{demographicData} (a data frame of demographic information).}

\item{params}{A character vector of parameter names to extract from \code{demographicData}. Default is \code{c('Age', 'Weight', 'Height')}.}
}
\value{
A list with \code{points}, a data frame of weighted support points, and \code{weights}, a vector of posterior probabilities.
}
\description{
Computes weighted support points using posterior probabilities and demographic parameters.
}
\examples{
res <- list(
theta = matrix(c(0.25,0.625,0.125,0.9375),ncol = 2),
w = c(0.7, 0.3),
PSI = matrix(runif(6),ncol=2),
demographicData = data.frame(Age = c(25, 30, 35),
Weight = c(70, 80, 90), Height = c(170, 175, 180))
)
getWeightedPoints(res)

}
