% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{vbeParameter}
\alias{vbeParameter}
\title{vbeParameter}
\usage{
vbeParameter(
  pathInReferenceSimulation,
  pathInTestSimulation,
  displayName = NULL,
  dimension,
  unit,
  lowerBound = 0.1,
  upperBound = 10,
  logIncrement = TRUE
)
}
\arguments{
\item{pathInReferenceSimulation}{Character. Path to the parameter in the reference simulation.}

\item{pathInTestSimulation}{Character. Path to the parameter in the test simulation.}

\item{displayName}{Character. Display name of the parameter (defaults to \code{pathInReferenceSimulation}).}

\item{dimension}{Character. Dimension of the parameter.}

\item{unit}{Character. Unit of the parameter.}

\item{lowerBound}{Numeric. Lower bound for the parameter (default: 1e-1).}

\item{upperBound}{Numeric. Upper bound for the parameter (default: 1e1).}

\item{logIncrement}{Logical. Whether the parameter is incremented log-arithmically (default: TRUE).}
}
\value{
A character vector of parameter paths in the test simulation.
}
\description{
Initialize a VBEParameter object.
}
